Worked on Issue #2
Normalized the markup for the input region heavily. Removed the
redundant markup. The redundant markup did not add value either in
presentation nor in structuring of information.
Normalized the application of style to elements and took help of
standard bootstrap styles. Additional styles were rationed. Application
of custom styles on need basis.